+++ colorer/src/shared/unicode/CharacterClass.cpp
@@ -219,7 +219,7 @@ void CharacterClass::addCategory(ECharCategory cat){
   };
 }
 void CharacterClass::addCategory(const String &cat){
-  for(int pos = 0; pos < ARRAY_SIZE(char_category_names); pos++){
+  for(size_t pos = 0; pos < ARRAY_SIZE(char_category_names); pos++){
     int ci;
     for(ci = 0; ci < cat.length() && cat[ci] == char_category_names[pos][ci]; ci++);
     if (ci == cat.length()) addCategory(ECharCategory(pos));
@@ -243,7 +243,7 @@ void CharacterClass::clearCategory(ECharCategory cat){
   };
 }
 void CharacterClass::clearCategory(const String &cat){
-  for(int pos = 0; pos < ARRAY_SIZE(char_category_names); pos++){
+  for(size_t pos = 0; pos < ARRAY_SIZE(char_category_names); pos++){
     int ci;
     for(ci = 0; ci < cat.length() && cat[ci] == char_category_names[pos][ci]; ci++);
     if (ci == cat.length()) clearCategory(ECharCategory(pos));
